Build the GNU multiple precision library the correct way.
The previous method of building had lots of problems

* Each variant would be the same object file multiple times which
  is very inefficient.
* Due to the above the linker also had to work much harder due
  to not linking against a library
* It is not portable as the build of gmp was using the config
  of a different machine rather than the host.

This commit instead builds and installs (into a temporary directory)
the gmp library and has the benchmarks link against it using the
newly added "out of tree" dependency support added to the fp-bench
infrastructure.

This unfortunately doesn't quite work properly yet because the
`gmp-6.1.1` directory is messed up (an in source build was
run in that directory). Subsequent commits will try to fix this.