Clang 3.4.2 hangs when trying to compile the `real/units/units-2.13/parse.tab.c` when optimizations are enabled.
The problem seems to be that the SROA pass gets stucks in an infinite
loop in Clang when compiling the `unitsparse` function. Unfortunately
``__attribute__((optnone))`` was not available in Clang 3.4.2 so
instead we add a little hack to the build system to force Clang to
build this particular benchmarks without optimizations regardless
of how CMAKE_BUILD_TYPE is set.